requirements.txt
biopython>=1.85
scipy
tqdm
pytest
pandas
networkx